@@ -1,11 +1,11 @@
 package com.group.libraryapp.dto.user.request
 
-import javax.validation.constraints.NotBlank
-import javax.validation.constraints.Size
+import jakarta.validation.constraints.NotBlank
+import jakarta.validation.constraints.Size
 
 data class UserSignInRequest(
     @field:NotBlank
     val email: String,
     @field:Size(min = 6, message = "at least 6 characters")
-    val password: String,
+    val password: String
 )
