@@ -6,14 +6,14 @@ import com.group.libraryapp.domain.user.User
 import com.group.libraryapp.exception.invalidEmailFail
 import com.group.libraryapp.type.company.CompanyStatus
 import com.group.libraryapp.type.company.EmployeeStatus
-import javax.persistence.CascadeType
-import javax.persistence.Entity
-import javax.persistence.EnumType
-import javax.persistence.Enumerated
-import javax.persistence.GeneratedValue
-import javax.persistence.GenerationType
-import javax.persistence.Id
-import javax.persistence.OneToMany
+import jakarta.persistence.CascadeType
+import jakarta.persistence.Entity
+import jakarta.persistence.EnumType
+import jakarta.persistence.Enumerated
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.OneToMany
 
 @Entity
 class Company(
