@@ -1,9 +1,9 @@
 package com.group.libraryapp.config
 
 import com.querydsl.jpa.impl.JPAQueryFactory
+import jakarta.persistence.EntityManager
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import javax.persistence.EntityManager
 
 @Configuration
 class QueryDSLConfig(
