@@ -7,13 +7,13 @@ import com.group.libraryapp.security.JWTTokenProvider
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.authentication.AuthenticationManager
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
 import org.springframework.security.config.http.SessionCreationPolicy
 import org.springframework.security.crypto.argon2.Argon2PasswordEncoder
 import org.springframework.security.crypto.password.PasswordEncoder
+import org.springframework.security.web.SecurityFilterChain
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
 import org.springframework.web.cors.CorsConfiguration
 import org.springframework.web.cors.CorsConfigurationSource
@@ -21,24 +21,25 @@ import org.springframework.web.cors.UrlBasedCorsConfigurationSource
 
 @Configuration
 @EnableWebSecurity
-@EnableGlobalMethodSecurity(securedEnabled = true)
-class SecurityConfig(jwtTokenProvider: JWTTokenProvider) : WebSecurityConfigurerAdapter() {
+class SecurityConfig(jwtTokenProvider: JWTTokenProvider) {
     private val accessTokenAuthenticationFilter = AccessTokenAuthenticationFilter(jwtTokenProvider)
 
-    override fun configure(http: HttpSecurity) {
-        http.httpBasic().disable()
-            .csrf().disable()
-            .cors().configurationSource(corsConfigurationSource()).and()
-            .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-            .and()
+    @Bean
+    fun configure(http: HttpSecurity): SecurityFilterChain {
+        http.httpBasic { it.disable() }
+            .csrf { it.disable() }
+            .cors { it.configurationSource(corsConfigurationSource()) }
+            .sessionManagement { it.sessionCreationPolicy(SessionCreationPolicy.STATELESS) }
             .addFilterBefore(logFilter(), UsernamePasswordAuthenticationFilter::class.java)
             .addFilterBefore(ExceptionHandlerFilter(), UsernamePasswordAuthenticationFilter::class.java)
             .addFilterAt(accessTokenAuthenticationFilter, UsernamePasswordAuthenticationFilter::class.java)
-            .headers().frameOptions().sameOrigin()
+            .headers { headers -> headers.frameOptions { it.sameOrigin() } }
+
+        return http.build()
     }
 
     @Bean
-    fun passwordEncoder(): PasswordEncoder = Argon2PasswordEncoder()
+    fun passwordEncoder(): PasswordEncoder = Argon2PasswordEncoder.defaultsForSpringSecurity_v5_8()
 
     @Bean
     fun corsConfigurationSource(): CorsConfigurationSource =
@@ -55,14 +56,19 @@ class SecurityConfig(jwtTokenProvider: JWTTokenProvider) : WebSecurityConfigurer
         }
 
     @Bean
-    override fun authenticationManagerBean(): AuthenticationManager =
-        super.authenticationManagerBean()
+    @Throws(Exception::class)
+    fun authenticationManager(authenticationConfiguration: AuthenticationConfiguration): AuthenticationManager? {
+        return authenticationConfiguration.authenticationManager
+    }
 
-    fun logFilter(): CustomLoggingFilter =
-        CustomLoggingFilter(hashSetOf("/ping"), hashSetOf("/sign-up")).apply {
-            setIncludeQueryString(true)
-            setIncludePayload(true)
-            setMaxPayloadLength(10240)
-            setIncludeHeaders(true)
-        }
+    @Bean
+    fun logFilter(): CustomLoggingFilter? {
+        val filter = CustomLoggingFilter()
+        filter.setIncludeQueryString(true)
+        filter.setIncludePayload(true)
+        filter.setMaxPayloadLength(1024 * 500)
+        filter.setIncludeHeaders(true)
+        filter.setAfterMessagePrefix("REQUEST DATA : ")
+        return filter
+    }
 }
